--- Gigs/Gigs/Models/Bearer.swift
@@ -1,13 +1,0 @@
-//
-//  Bearer.swift
-//  Gigs
-//
-//  Created by Ufuk Türközü on 15.01.20.
-//  Copyright © 2020 Ufuk Türközü. All rights reserved.
-//
-
-import Foundation
-
-struct Bearer: Codable {
-    let token: String
-}
