--- Gigs/Gigs/Models/User.swift
@@ -1,14 +1,0 @@
-//
-//  User.swift
-//  Gigs
-//
-//  Created by Ufuk Türközü on 15.01.20.
-//  Copyright © 2020 Ufuk Türközü. All rights reserved.
-//
-
-import Foundation
-
-struct User: Codable {
-    let username: String?
-    let password: String?
-}
