--- Gigs/Gigs/ViewControlle/GigDetailViewController.swift
@@ -1,37 +1,0 @@
-//
-//  GigDetailViewController.swift
-//  Gigs
-//
-//  Created by Ufuk Türközü on 16.01.20.
-//  Copyright © 2020 Ufuk Türközü. All rights reserved.
-//
-
-import UIKit
-
-class GigDetailViewController: UIViewController {
-
-    @IBOutlet weak var titleTextField: UITextField!
-    @IBOutlet weak var datePicker: UIDatePicker!
-    @IBOutlet weak var descriptionTV: UITextView!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-    @IBAction func saveTapped(_ sender: Any) {
-        
-    }
-    
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
