--- Gigs/Gigs/ViewControlle/GigsTableViewController.swift
@@ -1,68 +1,0 @@
-//
-//  GigsTableViewController.swift
-//  Gigs
-//
-//  Created by Ufuk Türközü on 15.01.20.
-//  Copyright © 2020 Ufuk Türközü. All rights reserved.
-//
-
-import UIKit
-
-class GigsTableViewController: UITableViewController {
-    
-    let gigController = GigController()
-    
-    var formatter: DateFormatter {
-        let df = DateFormatter()
-        df.dateStyle = .short
-        df.timeStyle = .none
-        return df
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    
-    }
-    
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        
-        if gigController.bearer == nil {
-            performSegue(withIdentifier: "LoginViewModalSegue", sender: self)
-        } // TODO: fetch gigs here
-    }
-
-    // MARK: - Table view data source
-
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        // #warning Incomplete implementation, return the number of rows
-        return gigController.gigs.count
-    }
-
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "GigCell", for: indexPath)
-
-        // Configure the cell...
-        let gig = gigController.gigs[indexPath.row]
-        cell.textLabel?.text = gig.title
-        cell.detailTextLabel?.text = formatter.string(from: gig.dueDate)
-
-        return cell
-    }
-
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-        if segue.identifier == "LoginViewModalSegue" {
-            if let destinationVC = segue.destination as? LoginViewController {
-                destinationVC.gigController = gigController
-            } else if segue.identifier == "" {
-
-            }
-        }
-    }
-
-}
