+++ IOS14/Gigs/Gigs/Gig.swift
@@ -2,14 +2,13 @@
 //  Gig.swift
 //  Gigs
 //
-//  Created by Ufuk Türközü on 16.01.20.
+//  Created by Ufuk Türközü on 13.02.20.
 //  Copyright © 2020 Ufuk Türközü. All rights reserved.
 //
 
 import Foundation
 
 struct Gig: Codable {
-    
     let title: String
     let dueDate: Date
     let description: String
