+++ IOS14/Gigs/Gigs/GigsTableViewController.swift
@@ -11,9 +11,12 @@ import UIKit
 class GigsTableViewController: UITableViewController {
     
     let gigController = GigController()
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
+    
+    var formatter: DateFormatter {
+        let df = DateFormatter()
+        df.dateStyle = .short
+        df.timeStyle = .none
+        return df
     }
     
     override func viewDidAppear(_ animated: Bool) {
@@ -21,21 +24,30 @@ class GigsTableViewController: UITableViewController {
         
         if gigController.bearer == nil {
             performSegue(withIdentifier: "LoginManualSegue", sender: self)
+        } else {
+            gigController.fetchAllGigs { result in
+                
+                DispatchQueue.main.async {
+                    self.tableView.reloadData()
+                }
+            }
         }
-        // TODO: fetch gigs here
     }
 
     // MARK: - Table view data source
 
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         // #warning Incomplete implementation, return the number of rows
-        return 0
+        return gigController.gigs.count
     }
 
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "GigCell", for: indexPath)
 
         // Configure the cell...
+        let gig = gigController.gigs[indexPath.row]
+        cell.textLabel?.text = gig.title
+        cell.detailTextLabel?.text = formatter.string(from: gig.dueDate)
 
         return cell
     }
@@ -45,8 +57,20 @@ class GigsTableViewController: UITableViewController {
     // In a storyboard-based application, you will often want to do a little preparation before navigation
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         if segue.identifier == "LoginManualSegue" {
-            if let destinationVC = segue.destination as? LoginViewController {
-                destinationVC.gigController = gigController
+            if let loginVC = segue.destination as? LoginViewController {
+                loginVC.gigController = gigController
+            }
+        } else if segue.identifier == "ShowGigDetailSegue" {
+            if let detailVC = segue.destination as? GigDetailViewController {
+                detailVC.gigController = gigController
+                if let indexPath = tableView.indexPathForSelectedRow {
+                    let gig = gigController.gigs[indexPath.row]
+                    detailVC.gig = gig
+                }
+            }
+        } else if segue.identifier == "AddGigSegue" {
+            if let detailVC = segue.destination as? GigDetailViewController {
+                detailVC.gigController = gigController
             }
         }
     }
